requirements.txt
pillow
shutil